requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
openapi-codec==1.3.2
packaging==23.1
protobuf==3.20.3
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
python-decouple==3.8
